smnist/requirements.txt
#numpy==1.24.4
numpy==1.23.0
#PyYAML==6.0
PyYAML==6.0.1
#scikit_learn==1.3.0
scikit_learn==1.2.2
scipy==1.9.1
spikingjelly==0.0.0.0.14
timm==0.8.17.dev0
#torch==2.0.0
torchaudio==2.0.0
wget==3.2
